Refactor validator events tabs
