+++ packages/webapp/components/validators/events/consolidations-tab.tsx
@@ -1,0 +1,8 @@
+'use client';
+
+import { EmptyStateTab } from './empty-state-tab';
+
+/** Renders the consolidations tab placeholder. */
+export function ConsolidationsTab() {
+  return <EmptyStateTab message="No consolidations" />;
+}
