+++ packages/webapp/components/validators/events/deposits-tab.tsx
@@ -1,0 +1,8 @@
+'use client';
+
+import { EmptyStateTab } from './empty-state-tab';
+
+/** Renders the deposits tab placeholder. */
+export function DepositsTab() {
+  return <EmptyStateTab message="No deposits" />;
+}
