+++ packages/webapp/components/validators/events/withdrawals-tab.tsx
@@ -1,0 +1,8 @@
+'use client';
+
+import { EmptyStateTab } from './empty-state-tab';
+
+/** Renders the withdrawals tab placeholder. */
+export function WithdrawalsTab() {
+  return <EmptyStateTab message="No withdrawals" />;
+}
